slight refactor: move stability tests into separate file
